refactor(nkigen-lite): extract per-op emitters + shared tile load/store helper
Split direct_lower.py (~1000 lines lighter) into focused per-op modules —
direct_lower_{fusion,gather,iota,topk}.py — and consolidate the elementwise
load/store path.

- Segmentation + elementwise emission move to direct_lower_fusion.py; gather,
  iota, and topk emitters get their own modules. direct_lower.py now dispatches
  to them.
- New load_input_tile / store_output_tile / canonical_layout in
  direct_lower_utils.py are the single elementwise tile load/store path,
  replacing the two near-identical inlined branches in _emit_ew_tile. The
  slice-view offset becomes an optional per-dim `offsets` descriptor composed
  into the DimSlices — the seam a future transpose/reshape view composes
  through.
- Dedup the hand-rolled batch-index unravel in broadcast/matmul/transpose
  emitters onto the shared unravel() helper.

Behavior-neutral: qwen3 MoE layer stays at 39968 nki ops / 10601 dma_copy.
All 435 non-HW tensor_ir tests pass.
